Fix DeleteCompany functionality

+++ src/main/java/command/DeleteCompany.java
@@ -1,6 +1,7 @@
 package command;
 
 import services.CompanyService;
+import utils.CheckCompanies;
 import view.View;
 
 import java.sql.SQLException;
@@ -9,6 +10,7 @@ public class DeleteCompany implements Command {
     public static final String DELETE_COMPANY = "delete company";
     private final View view;
     CompanyService companyService = new CompanyService();
+    CheckCompanies checkCompanies = new CheckCompanies();
 
     public DeleteCompany(View view) {
         this.view = view;
@@ -22,18 +24,22 @@ public class DeleteCompany implements Command {
     @Override
     public void execute() {
         int id;
-        while (true) {
-            try {
-                view.write("Please, enter company id to delete: ");
-                id = Integer.parseInt(view.read());
-                break;
-            } catch (NumberFormatException e) {
-                view.write("Invalid value. Use digits");
-            }
-        }
         try {
-            companyService.deleteCompany(id);
-            view.write("Company with id " + id + " successfully deleted");
+            while (true) {
+                try {
+                    view.write("Please, enter company id to delete: ");
+                    id = Integer.parseInt(view.read());
+                    if (checkCompanies.IsCompanyIdExists(id)) {
+                        companyService.deleteCompany(id);
+                        view.write("Company with id " + id + " successfully deleted");
+                        break;
+                    } else {
+                        System.out.println("Company id doesn't exists");
+                    }
+                } catch (NumberFormatException e) {
+                    view.write("Invalid value. Use digits");
+                }
+            }
         } catch (SQLException e) {
             e.getStackTrace();
         }
